強制アップデート用ダイアログクラスを移動
